+++ packages/iocuak/task/models/domain/TaskScope.ts
@@ -1,4 +1,5 @@
 export enum TaskScope {
   transient = 'transient',
+  request = 'request',
   singleton = 'singleton',
 }