+++ packages/iocuak/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - `TaskScope` enum is now exposed.
+- Update `TaskScope` with `request` scope.
 - Updated `inject` to throw an error if it's not used in a constructor parameter.